pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
     "anthropic>=0.69.0",
     "beautifulsoup4>=4.14.2",
     "chromadb>=1.1.0",
     "datasets==3.6.0",
     "feedparser>=6.0.12",
     "google-genai>=1.41.0",
     "google-generativeai>=0.8.5",
     "gradio>=5.47.2,<6.0",
     "ipykernel>=6.30.1",
     "ipywidgets>=8.1.7",
     "jupyter-dash>=0.4.2",
     "langchain>=0.3.27",
     "langchain-chroma>=0.2.6",
     "langchain-community>=0.3.30",
     "langchain-core>=0.3.76",
     "langchain-openai>=0.3.33",
     "langchain-text-splitters>=0.3.11",
     "litellm>=1.77.5",
     "matplotlib>=3.10.6",
     "nbformat>=5.10.4",
     "modal>=1.1.4",
     "numpy>=2.3.3",
     "ollama>=0.6.0",
     "openai>=1.109.1",
     "pandas>=2.3.3",
     "plotly>=6.3.0",
     "protobuf==3.20.2",
     "psutil>=7.1.0",
     "pydub>=0.25.1",
     "python-dotenv>=1.1.1",
     "requests>=2.32.5",
     "scikit-learn>=1.7.2",
     "scipy>=1.16.2",
     "sentence-transformers>=5.1.1",
     "setuptools>=80.9.0",
     "speedtest-cli>=2.1.3",
     "tiktoken>=0.11.0",
     "torch>=2.8.0",
     "tqdm>=4.67.1",
     "transformers>=4.56.2",
     "wandb>=0.22.1",
     "langchain-huggingface>=1.0.0",
     "langchain-ollama>=1.0.0",
     "langchain-anthropic>=1.0.1",
     "langchain-experimental>=0.0.42",
     "groq>=0.33.0",
     "xgboost>=3.1.1",
     "pypdf>=6.7.5",
     "fastapi[standard]>=0.135.1",
     "uvicorn>=0.41.0",
]
